refactor: split codebase into modular package structure
- Split monolithic handelsregister.py into focused modules:
  * constants.py: Global constants and mappings
  * exceptions.py: Custom exception classes
  * settings.py: Configuration management with pydantic-settings
  * models.py: Pydantic data models
  * cache.py: DiskCache-based caching implementation
  * parser.py: HTML parsing logic
  * client.py: Browser automation client
  * cli.py: Command-line interface
- Create handelsregister/__init__.py to expose public API
- Maintain backward compatibility with compatibility shim
- Update pyproject.toml entry point to new structure
- All tests passing (87 passed)
