+++ handelsregister/exceptions.py
@@ -1,0 +1,33 @@
+"""Exception classes for the Handelsregister package."""
+
+from typing import Optional
+
+
+class HandelsregisterError(Exception):
+    """Base exception for all Handelsregister errors."""
+    pass
+
+
+class NetworkError(HandelsregisterError):
+    """Raised when a network request fails."""
+    def __init__(self, message: str, original_error: Optional[Exception] = None):
+        super().__init__(message)
+        self.original_error = original_error
+
+
+class ParseError(HandelsregisterError):
+    """Raised when HTML parsing fails."""
+    def __init__(self, message: str, html_snippet: Optional[str] = None):
+        super().__init__(message)
+        self.html_snippet = html_snippet
+
+
+class FormError(HandelsregisterError):
+    """Raised when form interaction fails."""
+    pass
+
+
+class CacheError(HandelsregisterError):
+    """Raised when cache operations fail."""
+    pass
+
